requirements.txt
pandas
python-dotenv
scikit-learn
